Move event handlers to separate handlers.py
